一、定义array类；二、完成Tasks类；三、修改storage中的insert_task原型；四、调整文件结构，避免循环依赖

+++ storage.h
@@ -24,7 +24,7 @@ public:
     /// Get User fields
     User& user();
     /// Get Task array and write length into `len`
-    Task* tasks(int& len);
+    Task* tasks(uint64_t& len);
     /// Any task insertion MUST use insert_task method, to allocate
     /// new memory on file; otherwise SIGSEGV may be raised
     void insert_task(const Task& task);
